+++ usp-core-waap-virtual-patch/step1/verify.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+test -e /tmp/.backend_installed