+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 changes to the USP Killercoda scenarios using [common-changelog.org](https://common-changelog.org/) format.
 
-## [20xx-yy] - 20xx-mm-dd
+## [2025-03] - 2025-01-17
 
 ### Added
+
+- Added scenario `USP Core WAAP Virtual Patch`
 - Added scenario `USP Core WAAP OpenAPI validation`
 
 ## [2024-51] - 2024-12-17